+++ zcash_primitives/src/transaction/components/sprout.rs
@@ -17,6 +17,20 @@ pub struct Bundle {
     pub joinsplit_sig: [u8; 64],
 }
 
+impl Bundle {
+    /// The value balance for the bundle. When this is positive,
+    /// its value is added to the transparent value pool; when it
+    /// is negative, its value is subtracted from the transparent
+    /// value pool.
+    pub fn value_balance(&self) -> Option<Amount> {
+        self.joinsplits
+            .iter()
+            .try_fold(Amount::zero(), |total, js| {
+                js.value_balance().and_then(|b| total + b)
+            })
+    }
+}
+
 #[derive(Clone)]
 #[allow(clippy::upper_case_acronyms)]
 pub(crate) enum SproutProof {
@@ -171,5 +185,13 @@ impl JsDescription {
 
         writer.write_all(&self.ciphertexts[0])?;
         writer.write_all(&self.ciphertexts[1])
+    }
+
+    /// The value balance for the JoinSplit. When this is positive,
+    /// its value is added to the transparent value pool; when it
+    /// is negative, its value is subtracted from the transparent
+    /// value pool.
+    pub fn value_balance(&self) -> Option<Amount> {
+        self.vpub_new - self.vpub_old
     }
 }